requirements.txt
mysql-connector-python==8.2.0
python-dateutil==2.8.2
Flask==3.0.0
Flask-CORS==4.0.0
